+++ tests/WriteConcurrencyTests.cpp
@@ -1,6 +1,6 @@
 #include "tools/BuiltinToolRuntime.h"
-#include "tools/BuiltinTools/WriteFileTool.h"
-#include "tools/BuiltinTools/helpers/WorkspaceHelper.h"
+#include "tools/builtin/WriteFileTool.h"
+#include "tools/builtin/helpers/WorkspaceHelper.h"
 #include "tools/WriteCoordinator.h"
 
 #include <QtTest/QtTest>